Refactor navigator into separate module
