+++ site/.gitignore
@@ -11,6 +11,7 @@ node_modules
 dist
 dist-ssr
 *.local
+.DS_Store
 
 # Editor directories and files
 .vscode/*
