+++ site/src/navigator/NavigatorConfig.js
@@ -1,0 +1,142 @@
+export const tours = {
+  paris: {
+    title: "Paris Roads",
+    visibleTypes: [
+      "division_area",
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "segment",
+      "connector",
+      "building",
+      "building_part",
+    ],
+    theme: "transportation",
+    view: {
+      center: [2.3417, 48.8552],
+      zoom: 11.73,
+      pitch: 12,
+      bearing: 15.3,
+    },
+  },
+  nyc: {
+    title: "NYC Buildings",
+    visibleTypes: [
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "building",
+      "building_part",
+    ],
+    theme: "buildings",
+    view: {
+      center: [-73.99768, 40.75332],
+      zoom: 14.22,
+      pitch: 60,
+      bearing: 60.6,
+    },
+  },
+  world: {
+    title: "World Landcover",
+    visibleTypes: ["land", "land_cover", "land_use", "water", "infrastructure"],
+    theme: "base",
+    view: {
+      center: [10.3, 24.5],
+      zoom: 1.07,
+      bearing: 0,
+      pitch: 0,
+    },
+  },
+  london: {
+    title: "London Places",
+    visibleTypes: [
+      "division_area",
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "segment",
+      "building",
+      "building_part",
+      "place",
+    ],
+    theme: "places",
+    view: {
+      center: [-0.091217, 51.514511],
+      zoom: 16.02,
+      bearing: -24,
+      pitch: 50,
+    },
+  },
+  boston: {
+    title: "Boston Addresses",
+    visibleTypes: [
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "segment",
+      "building",
+      "building_part",
+      "address",
+    ],
+    theme: "addresses",
+    view: {
+      center: [-71.065192, 42.353714],
+      zoom: 15.94,
+      bearing: 0,
+      pitch: 52,
+    },
+  },
+  athens: {
+    title: "Athens Shipping Lines",
+    visibleTypes: [
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "segment",
+    ],
+    theme: "transportation",
+    view: {
+      center: [23.63247, 37.94203],
+      zoom: 14.83,
+      bearing: -142.3,
+      pitch: 60,
+    },
+  },
+  hongkong: {
+    title: "Hong Kong Port",
+    visibleTypes: [
+      "division_boundary",
+      "land",
+      "land_cover",
+      "land_use",
+      "water",
+      "infrastructure",
+      "segment",
+      "connector",
+      "building",
+      "building_part",
+    ],
+    theme: "buildings",
+    view: {
+      center: [114.15284, 22.2934],
+      zoom: 14.83,
+      bearing: 96.2,
+      pitch: 60,
+    },
+  },
+};
