+++ site/src/navigator/images/athens.png
