+++ site/src/navigator/images/london.png
