+++ site/src/navigator/images/nyc.png
