+++ site/src/navigator/images/world.png
